+++ src/ohana_agent/observation/events/__init__.py
@@ -1,7 +1,9 @@
 """Observation domain events."""
 
+from ohana_agent.observation.events.host_health_observed import HostHealthObserved
 from ohana_agent.observation.events.observation_published import ObservationPublished
 
 __all__ = [
+    "HostHealthObserved",
     "ObservationPublished",
 ]